Rereation
Completely redesigned the drag and drop system to use much less code and created a UI that is more pleasing to the eye. Main components are prefabbed and the UI redesign allows for much more modular creation of future levels.

Validation needs to be recreated still however that was buggy in the initial drag drop system as it could not detect more than one tag.